Create setup_conda.sh
This is an installation script that can automatically complete the stable version of anaconda3 2021.11. It is suitable for non-sudo installation of anaconda3 on Linux GPU servers. After the installation is complete, all conda installations will be completed in ~/anaconda3/ in your own user directory. Realize the function of one-click installation.